+++ cisco_acl/__init__.py
@@ -2,16 +2,27 @@
 
 from cisco_acl.ace import Ace
 from cisco_acl.ace_group import AceGroup
-from cisco_acl.acl import Acl, from_config
+from cisco_acl.acl import Acl
 from cisco_acl.address import Address
+from cisco_acl.config_to import config_to_ace, config_to_aceg
 from cisco_acl.port import Port
 from cisco_acl.protocol import Protocol
 from cisco_acl.remark import Remark
 
-__all__ = ["Ace", "AceGroup", "Acl", "Address", "Port", "Protocol", "Remark", "from_config"]
+__all__ = [
+    "Ace",
+    "AceGroup",
+    "Acl",
+    "Address",
+    "Port",
+    "Protocol",
+    "Remark",
+    "config_to_ace",
+    "config_to_aceg",
+]
 
-__version__ = "1.0.0"
-__date__ = "2022-07-16"
+__version__ = "1.1.0"
+__date__ = "2022-07-17"
 __title__ = "cisco-acl"
 
 __summary__ = "Python package to parse and manage Cisco ACL (Access Control List)"