+++ cisco_acl/parser/__init__.py
@@ -1,4 +1,0 @@
-"""CISCO config parser"""
-
-from cisco_acl.parser.parser_ios import ParserIOS
-from cisco_acl.parser.parser_nxos import ParserNXOS